requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
Babel==2.6.0
certifi==2018.10.15
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
daphne==2.2.3
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-annoying==0.10.4
django-bootstrap4==0.0.7
django-filters==0.2.1
django-hitcount==1.3.0
django-phonenumber-field==2.0.1
django-rest-auth==0.9.3
django-taggit==0.22.2
django-twilio==0.9.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
Markdown==3.0.1
msgpack==0.5.6
oauthlib==2.1.0
phonenumbers==8.9.16
Pillow==5.3.0
PyHamcrest==1.9.0
PyJWT==1.6.4
PySocks==1.6.8
python3-openid==3.1.0
pytz==2018.3
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
twilio==6.19.1
Twisted==18.9.0
txaio==18.8.1
typing==3.6.6
urllib3==1.24
zope.interface==4.6.0
